+++ app/src/main/java/com/example/frontpi4/dto/VendaDTO.java
@@ -4,23 +4,23 @@ import java.util.List;
 
 public class VendaDTO {
     private Long id;
-    private Long clienteId;
+    private int clienteId;
     private String data;
     private Double totalV;
     private List<ItemVendaDTO> itensVenda = null;
 
-    public VendaDTO(String data, Double totalV, Long clienteId, List<ItemVendaDTO> itensVenda) {
+    public VendaDTO(String data, Double totalV, int clienteId, List<ItemVendaDTO> itensVenda) {
         this.data = data;
         this.totalV = totalV;
         this.clienteId = clienteId;
         this.itensVenda = itensVenda;
     }
 
-    public Long getClienteId() {
+    public int getClienteId() {
         return clienteId;
     }
 
-    public void setClienteId(Long clienteId) {
+    public void setClienteId(int clienteId) {
         this.clienteId = clienteId;
     }
